Moved EquationSystemProblemOperator into its own file.
